+++ src/app/views/UserProfile.svelte
@@ -1,12 +1,15 @@
 <script lang="ts">
   import Input from "src/partials/Input.svelte"
+  import FlexColumn from "src/partials/FlexColumn.svelte"
+  import Card from "src/partials/Card.svelte"
   import ImageInput from "src/partials/ImageInput.svelte"
   import Textarea from "src/partials/Textarea.svelte"
   import Anchor from "src/partials/Anchor.svelte"
   import Footer from "src/partials/Footer.svelte"
   import Heading from "src/partials/Heading.svelte"
+  import Modal from "src/partials/Modal.svelte"
   import Field from "src/partials/Field.svelte"
-  import {pubkey, getProfile, publishProfile} from "src/engine"
+  import {env, pubkey, getProfile, publishProfile} from "src/engine"
   import {router} from "src/app/util/router"
 
   const nip05Url = "https://github.com/nostr-protocol/nips/blob/master/05.md"
@@ -14,12 +17,33 @@
   const pseudUrl =
     "https://www.coindesk.com/markets/2020/06/29/many-bitcoin-developers-are-choosing-to-use-pseudonyms-for-good-reason/"
 
-  const submit = () => {
-    publishProfile(values)
+  const closeModal = () => {
+    modal = null
+  }
+
+  const publishToPlatform = () => {
+    publishProfile(values, {forcePlatform: true})
+    router.pop()
     router.pop()
   }
 
+  const publishToNetwork = () => {
+    publishProfile(values, {forcePlatform: false})
+    router.pop()
+    router.pop()
+  }
+
+  const submit = () => {
+    if ($env.PLATFORM_RELAYS.length === 0) {
+      publishToNetwork()
+    } else {
+      modal = 'select-scope'
+    }
+  }
+
   const values = {...getProfile($pubkey)}
+
+  let modal
 
   document.title = "Profile"
 </script>
@@ -85,3 +109,42 @@
     <Anchor grow button tag="button" type="submit">Save</Anchor>
   </Footer>
 </form>
+
+{#if modal === 'select-scope'}
+  <Modal onEscape={closeModal}>
+    <div class="mb-4 flex flex-col items-center justify-center">
+      <Heading>Update Profile</Heading>
+      <p>Where would you like to publish your profile?</p>
+    </div>
+    <Card interactive on:click={publishToNetwork}>
+      <FlexColumn>
+        <div class="flex items-center justify-between">
+          <p class="flex items-center gap-4 text-xl">
+            <i class="fa fa-share-nodes text-neutral-600" />
+            <strong>The wider nostr network</strong>
+          </p>
+          <i class="fa fa-arrow-right" />
+        </div>
+        <p>
+          Publishing your profile to the wider nostr network will allow anyone to see it.
+          Use this if you plan to use other clients or relay selections.
+        </p>
+      </FlexColumn>
+    </Card>
+    <Card interactive on:click={publishToPlatform}>
+      <FlexColumn>
+        <div class="flex items-center justify-between">
+          <p class="flex items-center gap-4 text-xl">
+            <i class="fa fa-thumbtack text-neutral-600" />
+            <strong>Just this instance</strong>
+          </p>
+          <i class="fa fa-arrow-right" />
+        </div>
+        <p>
+          Publish your profile just to the relays configured on this instance if you prefer.
+          Be aware that how private this is depends on how the instance operator has set things up.
+        </p>
+      </FlexColumn>
+    </Card>
+  </Modal>
+{/if}